# Conflicts:
#	HandheldCompanion/Managers/OSDManager.cs
#	HandheldCompanion/Managers/Overlay/OverlayColors.cs
#	HandheldCompanion/Managers/Overlay/Strategy/CustomStrategy.cs
#	HandheldCompanion/Managers/Overlay/Strategy/DisabledStrategy.cs
#	HandheldCompanion/Managers/Overlay/Strategy/ExtendedStrategy.cs
#	HandheldCompanion/Managers/Overlay/Strategy/ExternalStrategy.cs
#	HandheldCompanion/Managers/Overlay/Strategy/FullStrategy.cs
#	HandheldCompanion/Managers/Overlay/Strategy/MinimalStrategy.cs
#	HandheldCompanion/Managers/Overlay/Widget/BatteryWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/CpuWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/FpsWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/GpuWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/RamWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/TimeWidget.cs
#	HandheldCompanion/Managers/Overlay/Widget/VramWidget.cs
#	HandheldCompanion/Managers/Overlay/WidgetFactory.cs
#	HandheldCompanion/Managers/OverlayManager.cs

+++ HandheldCompanion/Views/QuickPages/QuickKeyboardPage.xaml.cs
@@ -286,8 +286,8 @@ namespace HandheldCompanion.Views.QuickPages
                 }
                 else if (b.Tag is ValueTuple<char, uint> ct)
                 {
-                    seq.Add(MakeScan((uint)ct.Item1, ct.Item2, false));                         // Key down
-                    seq.Add(MakeScan((uint)ct.Item1, ct.Item2, true));                          // Key up
+                    seq.Add(MakeScan(ct.Item1, ct.Item2, false));                         // Key down
+                    seq.Add(MakeScan(ct.Item1, ct.Item2, true));                          // Key up
                 }
             }
